# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelZone AI/SOAR",
    "description": "Standalone investigation and human-approved SOAR platform. Dry-run results never represent live containment.",
    "version": "0.30.0-rc.1"
  },
  "paths": {
...
    "/v1/playbooks": {
      "get": {
        "tags": [
          "Playbooks and policy"
        ],
        "summary": "Listing",
        "operationId": "listing_v1_playbooks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybookPage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/playbooks/{playbook_id}": {
      "get": {
        "tags": [
          "Playbooks and policy"
        ],
        "summary": "Detail",
        "operationId": "detail_v1_playbooks__playbook_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "playbook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playbook Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playbook"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/v1/policy": {
      "get": {
        "tags": [
          "Playbooks and policy"
        ],
        "summary": "Policy",
        "operationId": "policy_v1_policy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Condition": {
        "properties": {
          "field": {
            "type": "string",
            "enum": [
              "action_type",
              "target",
              "parameters.lab_only"
            ],
            "title": "Field"
          },
          "equals": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Equals"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "equals"
        ],
        "title": "Condition"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ErrorIssue"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ErrorDetail"
      },
      "ErrorIssue": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "loc",
          "type"
        ],
        "title": "ErrorIssue"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "Playbook": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Name"
          },
          "version": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Version"
          },
          "action_type": {
            "type": "string",
            "enum": [
              "BLOCK_IP",
              "SUSPEND_TEST_PROCESS"
            ],
            "title": "Action Type"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PlaybookStep"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 5,
            "title": "Steps"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "action_type",
          "steps"
        ],
        "title": "Playbook"
      },
      "PlaybookPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Playbook"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PlaybookPage"
      },
      "PlaybookStep": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "validate_target",
              "human_approval",
              "windows_firewall_block",
              "suspend_test_process",
              "verify_block",
              "verify_process",
              "rollback_block",
              "rollback_process",
              "branch"
            ],
            "title": "Action"
          },
          "depends_on": {
            "items": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Depends On"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          },
          "condition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Condition"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "action"
        ],
        "title": "PlaybookStep"
      },
      "PolicyView": {
        "properties": {
          "protected_assets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Protected Assets"
          },
          "protected_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Protected Roles"
          },
          "protected_networks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Protected Networks"
          },
          "lab_asset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Lab Asset Ids"
          },
          "managed_by": {
            "type": "string",
            "title": "Managed By"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "protected_assets",
          "protected_roles",
          "protected_networks",
          "lab_asset_ids",
          "managed_by"
        ],
        "title": "PolicyView"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}